langgraph_bot/pyproject.toml
[project]
name = "langgraph-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.4"
dependencies = [
    "arxiv>=2.4.0",
    "huggingface-hub>=1.2.3",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "langsmith>=0.5.2",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
]
